week_6/src/homework/requirements.txt
avro
requests
faust
fastavro
kafka-python==1.4.6
confluent_kafka